requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-countries==7.6.1
django-debug-toolbar==5.1.0
django-filter==25.1
django-guardian==2.4.0
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.15.2
dnspython==2.7.0
drf-nested-routers==0.94.1
email_validator==2.2.0
environ==1.0
gunicorn==23.0.0
idna==3.10
jwt==1.3.1
Markdown==3.7
packaging==24.2
phonenumbers==9.0.1
psycopg2-binary==2.9.10
pycparser==2.22
python-decouple==3.8
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
